add more methods to comment repo and use them in comment service
